requirements.txt
# ===========================
# KhelBot Dependencies
# ===========================

python-telegram-bot==20.7
google-generativeai==0.8.3
httpx==0.27.0
supabase==2.9.1
python-dotenv==1.0.1
apscheduler==3.10.4

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
